Added dll.c containing DllMain for automatic WSAStartup, modified hiredis.c to contain explicit cast and remove compiler warning, fixed LinkDLL in vcproj, modified net.c to get winsock error messages with win32err.h, corrected socket timeout code, changed example.c to take host argument, removed MSVC rubbish

+++ hiredis/hiredis.c
@@ -409,7 +409,7 @@ static int processBulkItem(redisReader *r) {
     if (s != NULL) {
         p = r->buf+r->pos;
         bytelen = s-(r->buf+r->pos)+2; /* include \r\n */
-        len = readLongLong(p);
+        len = (long)readLongLong(p);
 
         if (len < 0) {
             /* The nil object can always be created. */
@@ -464,7 +464,7 @@ static int processMultiBulkItem(redisReader *r) {
     }
 
     if ((p = readLine(r,NULL)) != NULL) {
-        elements = readLongLong(p);
+        elements = (long)readLongLong(p);
         root = (r->ridx == 0);
 
         if (elements == -1) {
